store: bring backup.go back into store.go
Issue: https://github.com/adamdecaf/cert-manage/issues/174
